Adjustments so that Test Errors could be Clearer

+++ rhsoft/src/main/java/br/com/gabriel/rhsoft/controllers/DepartmentsController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
+import br.com.gabriel.rhsoft.NoExposedCompanyException;
 import br.com.gabriel.rhsoft.daos.DepartmentsDAO;
 import br.com.gabriel.rhsoft.models.Company;
 import br.com.gabriel.rhsoft.models.Department;
@@ -24,12 +25,21 @@ public class DepartmentsController {
     //#region Forms
 
     @RequestMapping(value = "/form", name = "departmentForm")
-    public String departmentForm() {
+    public String departmentForm(HttpSession session) {
+        if(getExposedCompany(session) == null){
+            throw new NoExposedCompanyException();
+        }
+
         return "departments/departmentsForm";
     }
 
     @RequestMapping(value="/edit", name = "departmentEditForm", method = RequestMethod.POST)
-    public ModelAndView departmentEditForm(Integer id){
+    public ModelAndView departmentEditForm(Integer id, HttpSession session) {
+
+        if(getExposedCompany(session) == null){
+            throw new NoExposedCompanyException();
+        }
+
         ModelAndView modelAndView = new ModelAndView("departments/departmentsEditForm");
 
         Department department = departmentsDAO.findById(id);
@@ -46,7 +56,7 @@ public class DepartmentsController {
     public String createDepartment(Department department, HttpSession session) {
 
         if(getExposedCompany(session) == null){
-            throw new NullPointerException();
+            throw new NoExposedCompanyException();
         }
 
         department.setCompany(getExposedCompany(session));        
@@ -58,6 +68,12 @@ public class DepartmentsController {
 
     @RequestMapping(name = "removeDepartment", value = "/remove/{id}", method = RequestMethod.POST)
     public String removeDepartment(@PathVariable(value = "id") Integer id,  HttpSession session) {
+
+        if(getExposedCompany(session) == null){
+            throw new NoExposedCompanyException();
+        }
+
+
         Company company = getExposedCompany(session);
         departmentsDAO.deleteById(id, company.getId());
 
@@ -66,14 +82,22 @@ public class DepartmentsController {
 
     @RequestMapping(value = "/editDepartment", name = "editDepartment", method = RequestMethod.POST)
     public String editDepartment(Department department, HttpSession session) {
+
+        if(getExposedCompany(session) == null){
+            throw new NoExposedCompanyException("No Exposed Company");
+        }
         
         departmentsDAO.editDepartment(department, getExposedCompany(session));
-
+        
         return("redirect:" + urlUpdateExposed(session));
     }
 
     @RequestMapping(value = "/detail", name = "detailDepartment")
-    public ModelAndView detailDepartment(Integer id){
+    public ModelAndView detailDepartment(Integer id, HttpSession session) {
+        if(getExposedCompany(session) == null){
+            throw new NoExposedCompanyException("No Exposed Company");
+        }
+
         ModelAndView modelAndView = new ModelAndView("/departments/detail");
         Department department = departmentsDAO.findById(id);
 
@@ -112,8 +136,13 @@ public class DepartmentsController {
     }
 
     private Company getExposedCompany(HttpSession session){
-        Company company = (Company)session.getAttribute("exposedCompany");
-        return company;
+        try{
+            Company company = (Company)session.getAttribute("exposedCompany");
+            return company;
+        }catch(Exception e){
+            return null;
+        }
+        
     }
